requirements.txt
requests[socks]
beautifulsoup4
phonenumbers
geopy
exif
googlesearch-python
python-whois
